Implement a --protected-region-size flag to force a specific protected region size.
Statically-linked binaries need to know (at compilation time, before elf2tab runs) where they are located. Application code begins after the protected region size of a Tock binary, which is at least the size of the header. --protected-region-size allows application Makefiles to specify how large that region is. If the header does not fit in the region, elf2tab will return an error.