Update elf crate to v0.7.1
The changes from 0.0.10 -> 0.7.1 represent implementing more complete ELF parsing
functionality, performance improvements for zero-alloc and zero-copy parsing methods,
fuzz testing of the interface, and various bug fixes.

The changes seen in this patch aim to fix it up for the new interface.

The elf crate now supports two parsing modes:
1. Parsing lazily out of a &[u8] of the whole file data (the route I chose here)
2. Parsing lazily out of a std::io::<Read + Seek> via elf::ElfStream

It seemed to me like this tool inspects a lot of the file data anyway, so the former
is likely faster as opposed to the latter which allows lazy i/o as the parser seeks
around the file parsing things as requested.