Generate segment padding from ELF-internal offsets
Previously, `segment.p_addr` was used to generate offsets. Given a
specific ordering of ELF segments and certain target memory layouts,
this has been observed to insert very large paddings (~256MB), which
further break internal offsets in the binary. Instead of the target's
addresses, we should be using ELF-file internal offsets to infer
padding between segments (which means that the TBF-internal offsets
should be equivalent to those of the ELF file itself).