Warn on encountering section padding underflow
This can happen when iterating over sections which are not placed in
physical address order in the ELF file. When this occurs, elf2tab has
a chance of producing an incorrect TBF file, which we should warn
users about.

It can be an indication of an incorrect / broken ELF file, although
placing sections in an arbitrary order within an ELF file may well be
allowed. This change is deliberately conservative by issuing a warning
instead of exiting with an error, to avoid breaking anyone's
toolchain. It may be reasonable to sort sections by their physical
address in the future, if this is confirmed to be correct behavior.