segments need to start with a section
The linker can put additional data in a segment before a section, but
that violates assumptions that our linker files are making. elf2tab
requires that data exist within sections, and expects that segments
start with sections.

Most of the time this requirement is upheld by the linker. However, in
the case of VERY small apps, the linker tries to optimize by packing in
more in the first segment. This makes the segment start where a section
does not.

This patch adds logic to check for this discrepency and adjust the
segment to remove this (essentially) padding at the beginning of the
segment. The rest of elf2tab can then operate as expected.