use privatekey.pk8 instead
It seems that different OpenSSL versions create different .der types,
and the ring library we are using doesn't necessarily support them all.
Try using .pk8 instead.

While we are at it, we do not need a separate library for public keys
and we can get the public key from the private key pk8 file.