Use reserved space footer for padding
elf2tab supports two types of padding: 1) creating a reserved space
footer or 2) just adding bytes to the binary. The need to do both is
because a reserved space footer can only pad 8 or more bytes.

This patch fixes an issue where the padding needed is less than 8, but,
there is already going to be a reserved footer. In that case, we were
adding the padding twice (once in the reserved footer and again as just
extra bytes).

This issue only arises if you ask for a certain amount of footer space
when running elf2tab.